Broke everything but added valid payment check, refactored some borrowing (nothing works)

+++ src/conversions.rs
@@ -1,3 +1,3 @@
-pub fn btc_to_satoshi(amount: f32) -> f32 {
+pub fn btc_to_satoshi(amount: f64) -> f64 {
     amount * 100000000.0
 }
